Simplify API and implementation
API is now more compatible with cats, as it directly returns a
`ValidatedNel`, containing the error type.

The DSL has also been cleaned up to that using the default error type
(`ConstraintViolation`) is much simpler, but providing a custom error
type is still easy, and just as powerful.

Currently, `ConstraintViolation` are pretty dumb. The next step is to
have them pull in some meta-data about the violation; specifically the
path to the value being validated (if this was a nested value) and the
type of the subject.