Add instances for various cats type-classes
`Validator` now has instances for `Invariant`, `Cartesian` and `Monoid`.

All of these instances have been verified using discipline to check the
laws. This is tricky, since there's no `Eq` instance for `Validator`,
which discipline requires in order to verify the laws hold.

As a work-around, the tests use a fudged `Eq` instance that abuse
ScalaCheck to verify that the two inputs are equal.