Add tracking the path of an object graph
When validating a large object graph, it can be confusing to know
exactly which nodes a given error refers to. Textual error messages
alone aren't enough. This is an even greater problem for graphs with the
same data-type repeated throughout.

Now, `ConstraintViolation`s track the path through a `List[String]` of
names. These names can be provided by the validator constraints at use-site,
however it's customary for these to reflect the name of the
property/field being validated.