Add syntax sugar
An implicit class has been added to give any object the `validated`
method, which takes care of finding the implicit `Validator` and
validating against it.

`Validator.apply[A]` has been re-purposed as a means to get the implicit
`Validator` for the given type, using `ConstraintViolation` as the error
type.

The old constructor, `Validator.apply[A]` has been renamed
`Validator.of[A]` to enable the above change.